% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bson.R
\name{create_bson_oid}
\alias{create_bson_oid}
\title{Create BSON Object IDs}
\usage{
create_bson_oid(n = 1)
}
\arguments{
\item{n}{The number of OIDs to create.}
}
\value{
A vector of class \code{"bson_oid"}.
}
\description{
When it is necessary to keep track of BSON OIDs, you may want to create them
manually.
}
\examples{
df <- data.frame(`_id` = I(create_bson_oid(10)),
                 x = 1:10, check.names = FALSE)
\dontrun{

# Insert this data into Mongo.
m <- mongo("test_oid", verbose = FALSE)
m$drop()
m$insert(df)

# The OIDs of the output should match.
m$find(fields = "{}")
}

}
