% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoclient.R
\name{mongo_client}
\alias{mongo_client}
\title{MongoDB client}
\usage{
mongo_client(url = "mongodb://localhost", verbose = FALSE,
  options = ssl_options())
}
\arguments{
\item{url}{address of the mongodb server in mongo connection string
\href{http://docs.mongodb.org/manual/reference/connection-string}{URI format}}

\item{verbose}{emit some more output}

\item{options}{additional connection options such as SSL keys/certs.}
}
\value{
Upon success returns a pointer to database connection instance.
}
\description{
Connect to a MongoDB database. Returns a \link{mongo_client} connection object with
methods listed below.
}
\section{Methods}{

\describe{
\item{\code{use(collection, db)}}{Creates a \link{mongo} object representing a collection in a database with the active connection}
\item{\code{server_status()}}{Returns the server status.}
}
}

\examples{
# Connect to mongolabs
con <- mongo_client(url = "mongodb://readwrite:test@ds043942.mongolab.com:43942/jeroen_test")
con$default_database()
coll <- con$use('testcollection', 'jeroen_test')
coll$find()
rm(con)
gc()
}
\references{
Jeroen Ooms (2014). The \code{jsonlite} Package: A Practical and Consistent Mapping Between JSON Data and \R{} Objects. \emph{arXiv:1403.2805}. \url{http://arxiv.org/abs/1403.2805}
}
