% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{oid_to_timestamp}
\alias{oid_to_timestamp}
\title{Get OID date}
\usage{
oid_to_timestamp(oid)
}
\arguments{
\item{oid}{string or raw value with document oid}
}
\description{
The initial 4 bytes of a MongoDB OID contain a timestamp value, representing
the ObjectId creation, measured in seconds since the Unix epoch.
}
\examples{
oid_to_timestamp('5349b4ddd2781d08c09890f3')
}
