% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/object.R
\name{mongo}
\alias{mongo}
\title{MongoDB client}
\usage{
mongo(collection = "test", db = "test", url = "mongodb://localhost")
}
\arguments{
\item{collection}{name of collection}

\item{db}{name of database}

\item{url}{address of the mongodb server}
}
\description{
Initiate a new MongoDB collection.
}
\section{Methods}{

\describe{
  \item{\code{count(query = '{}')}}{Count the number of records for a given query. If no query is specified, counts all records.}
  \item{\code{drop()}}{Delete entire collection and all data.}
  \item{\code{find(query = '{}', fields = '{"_id" : 0}', skip = 0, limit = 0, handler = NULL, pagesize = 1000, verbose = TRUE)}}{Retrieve 'fields' from all records matching a query. If handler = NULL, data will returned as a single data frame.}
  \item{\code{index(add = NULL, remove = NULL)}}{find, add or remove indexes from the collection. Always returns a data frame with current indexes.}
  \item{\code{info()}}{Returns collection name and statistics (if available).}
  \item{\code{insert(data, pagesize = 1000, verbose = TRUE)}}{Insert a data frame into the collection.}
  \item{\code{remove(query = "{}", multiple = TRUE)}}{Remove record(s) from the collection.}
  \item{\code{rename(name, db = "test")}}{Change the name or database of a collection. Changing name is cheap, but changing database is expensive.}
}
}
\examples{
\dontrun{dplyr example
library(nycflights13)

# Insert some data
con <- mongo(collection = "myflights")
if(con$count() > 0) con$drop()
con$insert(flights)

# Basic operations
con$count('{"month":1, "day":1}')
mydata <- con$find(query = '{"month":1, "day":1}')
}
}

