% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/object.R
\name{mongo}
\alias{mongo}
\title{MongoDB client}
\usage{
mongo(url = "mongodb://localhost", db = "test", collection = "test")
}
\arguments{
\item{url}{address of the mongodb server}

\item{db}{name of database}

\item{collection}{name of collection}
}
\description{
Initiate a new MongoDB client.
}
\section{Methods}{

\describe{
  \item{\code{count(query = '{}')}}{Count the number of records for a given query. If no query is specified, counts all records.}
  \item{\code{drop()}}{Delete all records from the collection.}
}
}
\examples{
\dontrun{dplyr example
library(nycflights13)

# Insert some data
con <- mongo(collection = "myflights")
if(con$count() > 0) con$drop()
con$insert(flights)

# Basic operations
con$count('{"month":1, "day":1}')
mydata <- con$find(query = '{"month":1, "day":1}')
}
}

