% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{read_bson}
\alias{read_bson}
\title{Standalone BSON reader}
\usage{
read_bson(file, as_json = FALSE, verbose = interactive())
}
\arguments{
\item{file}{path or url to a bson file}

\item{as_json}{read data into json strings instead of R lists.}

\item{verbose}{print some progress output while reading}
}
\description{
Utility to read BSON data into R without MongoDB. Useful to read a dump from
from a \code{mongoexport} dump if it fits in memory. This utility does not attempt
to convert data into a data.frame: the output is a vector with length of the
total number of documents in the bson collection.
}
\details{
Alternatively, to import a bson dump into your local mongodb server, use the
\link[=mongo]{mongo$import} function instead. This requires little memory and once
data is in mongo you can easily query it.
}
\examples{
diamonds <- read_bson("https://jeroen.github.io/data/diamonds.bson")
length(diamonds)
}
