% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{read_bson}
\alias{read_bson}
\title{Standalone BSON reader}
\usage{
read_bson(con, as_json = FALSE, verbose = TRUE)
}
\arguments{
\item{con}{either a path to a file, a url, or a a connection object}

\item{as_json}{return data as json strings instead of R lists}

\item{verbose}{print some output as we read}
}
\description{
Utility to parse BSON files into R without using MongoDB. This is useful
to read data from a \code{mongoexport} dump without needing local mongodb server.
This reads all data in memory. To import a bson dump into your local mongodb
server, use the \link[=mongo]{mongo$import} function instead.
}
\examples{
diamonds <- read_bson("http://jeroen.github.io/data/diamonds.bson")
}
