% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{read_bson}
\alias{read_bson}
\title{Standalone BSON reader}
\usage{
read_bson(con, as_json = FALSE, verbose = interactive())
}
\arguments{
\item{con}{either a path to a file, a url, or a a connection object}

\item{as_json}{return data as json strings instead of R lists}

\item{verbose}{print some output as we read}
}
\value{
list with either data objects or json strings
}
\description{
Utility to parse BSON data into R without using MongoDB. Useful to read data
from a \code{mongoexport} dump without mongodb if it fits in memory.
}
\details{
To import a bson dump into a local mongodb server, use the \link[=mongo]{mongo$import}
function instead. This requires less memory and once data is in mongo you can
easily query it.
}
\examples{
diamonds <- read_bson("http://jeroen.github.io/data/diamonds.bson")
}
